refactor: new TrainServiceManagerAdapter for Train Input Port
